+++ lib/nretnil/utilities/array.rb
@@ -17,4 +17,16 @@ class Array
     end
     false
   end
+
+  ####
+  # Name: json_pretty
+  # Description: takes an Array object and returns json in a readable format
+  # Arguments: data - hash - object to be transformed
+  # Response: string
+  ####
+  def json_pretty
+    JSON.pretty_generate(self)
+  rescue NoMethodError
+    self
+  end
 end