+++ lib/nretnil/utilities/hash.rb
@@ -8,11 +8,13 @@ require 'json'
 class Hash
   ####
   # Name: json_pretty
-  # Description: takes a hash object and returns json is readable format
+  # Description: takes a hash object and returns json in a readable format
   # Arguments: data - hash - object to be transformed
   # Response: string
   ####
   def json_pretty
     JSON.pretty_generate(self)
+  rescue NoMethodError
+    self
   end
 end